user_roles/urls.py
from rest_framework.routers import DefaultRouter
from .views import UserViewSet, RoleViewSet
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

router = DefaultRouter()
router.register(r'users', UserViewSet)
router.register(r'roles', RoleViewSet)

urlpatterns = router.urls
