requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.6.2
anywidget==0.2.3
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
channels==3.0.2
channels-redis==4.0.0
chardet==5.2.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
constantly==15.1.0
contourpy==1.0.7
cookiecutter==2.5.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
cssselect2==0.7.0
cycler==0.11.0
daphne==3.0.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.1
django-allauth==0.54.0
django-annoying==0.10.6
django-appconf==1.0.5
django-background-tasks==1.2.5
django-channels==0.7.0
django-chartjs==2.3.0
django-ckeditor==6.6.1
django-client-side-image-cropping==0.1.10
django-compat==1.0.15
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-decouple==2.1
django-extensions==3.2.1
django-filter==23.2
django-formtools==2.4.1
django-image-cropping==1.7
django-imagekit==4.1.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-modeltranslation==0.18.9
django-multiselectfield==0.1.12
django-otp==1.2.1
django-otp-twilio==1.0.3
django-phonenumber-field==6.4.0
django-redis==5.2.0
django-responsive2==0.1.3
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-select2==8.1.1
django-storages==1.13.2
django-two-factor-auth==1.15.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
easy-thumbnails==2.8.5
elastic-transport==8.4.0
elasticsearch==8.6.0
executing==1.2.0
fastapi==0.108.0
filelock==3.12.0
Flask==2.2.2
Flask-RESTful==0.3.9
flet==0.19.0
flet-core==0.19.0
flet-runtime==0.19.0
fonttools==4.39.3
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
image==1.5.33
incremental==22.10.0
ipykernel==6.22.0
ipython==8.13.1
ipywidgets==8.0.6
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lxml==4.9.2
mapwidget==0.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
msgpack==1.0.4
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.24.2
oauthlib==3.2.2
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==23.2
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
pipenv==2023.5.19
platformdirs==3.5.0
plotly==5.14.0
prompt-toolkit==3.0.38
psutil==5.9.5
psygnal==0.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
pycryptodome==3.17
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.15.1
pyHanko==0.17.0
pyhanko-certvalidator==0.20.1
PyJWT==2.7.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pypdf==3.5.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.3
python-dotenv==1.0.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.4.2
ratelim==0.1.6
razorpay==1.3.0
redis==4.5.0
repath==0.9.0
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.7.0
service-identity==21.1.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.3.1
stack-data==0.6.2
starlette==0.32.0.post1
stripe==5.4.0
svglib==1.5.1
tenacity==8.2.2
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.1
tornado==6.3.1
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
uvicorn==0.25.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==2.2.2
whitenoise==6.4.0
widgetsnbextension==4.0.7
xhtml2pdf==0.2.9
xyzservices==2023.2.0
zope.interface==5.5.2
